Sync OpenSSL: cms: Create test for for purpose verification in cms application
From OpenSSL commit 19914fec9bac08ca7c7917eddc1b7d1dba67e4a7
https://github.com/openssl/openssl/pull/18567

The tests only cover the correct handling of the codesigning purpose in the certificates
in the context of the cms command line tool.
The interpretation of the certificate purpose is tested in the context of the "verify"
app. The correct handling of the cms objects is tested by other tests in 80-test_cms.t.