Sync OpenSSL: Transfer the functionality from ssl3_read_n to the new record layer From OpenSSL commit e2d5742b1460c45bf39094ea08e4e85a8f507ea8 https://github.com/openssl/openssl/pull/18132
This transfers the low level function ssl3_read_n to the new record layer.
We temporarily make the read_n function a top level record layer function.
Eventually, in later commits in this refactor, we will remove it as a top
level function and it will just be called from read_record.

But some CI cannot pass here, which will be fixed by following commits.