Sync OpenSSL: Add support for moving data from one epoch to the next From OpenSSL commit 359affdead3af497f1673204c5c34061d28dfa7b
Sometimes data read by a record layer in one epoch is actually intended for
the next epoch. For example in a TLS with read_ahead, the read_ahead data
could contain a KeyUpdate message followed by application data encrypted
with new keys. Therefore we implement a mechanism for passing this data
across the epochs.