Sync OpenSSL: Update the tls13encryptiontest for new read record layer
From OpenSSL commit a16f9d3366a4b4e8c8014bbf39b86baaf1a04047

The tls13encryption is an internal test that reaches inside libssl
to test encryption/decryption of records. It needs to be amended for the
new code structure so that it is testing the equivalent things as before.