Sync OpenSSL: Fix the return type for the rlayer_skip_early_data callback
From OpenSSL commit e921882d57201e14cc6a48765b2281065d6f5c65

There was a copy & paste error in the definition of the
rlayer_skip_early_data callback. The return type is supposed to
be "int" but it was defined as a pointer type. This was causing
test failures on some platforms.

Fixes #19037